// src/types/film.ts
export interface IFilm {
    title: string
    episode_id: number
    opening_crawl: string
    director: string
    producer: string
    release_date: string
    species: string[]
    starships: string[]
    vehicles: string[]
    characters: string[]
    planets: string[]
    url: string
    created: string
    edited: string
}

export interface IFilmsResponse {
    count: number
    next: string
    previous: string
    results: IFilm[]
}